requirements.txt
torch
torchvision
pandas
scikit-learn
matplotlib
seaborn
pillow
numpy